Separate modules of examples/dialogue_bot
