+++ examples/dialogue_bot/src/main.rs
@@ -20,126 +20,14 @@
 #[macro_use]
 extern crate frunk;
 
-use parse_display::Display;
-
-use favourite_music::FavouriteMusic;
-use teloxide::prelude::*;
-
 mod favourite_music;
+mod states;
+mod transitions;
 
-// Dialogue states.
+use states::*;
+use transitions::*;
 
-struct StartState;
-
-struct ReceiveFullNameState {
-    rest: StartState,
-}
-
-struct ReceiveAgeState {
-    rest: ReceiveFullNameState,
-    full_name: String,
-}
-
-struct ReceiveFavouriteMusicState {
-    rest: ReceiveAgeState,
-    age: u8,
-}
-
-#[derive(Display)]
-#[display(
-    "Your full name: {rest.rest.full_name}, your age: {rest.age}, your \
-     favourite music: {favourite_music}"
-)]
-struct ExitState {
-    rest: ReceiveFavouriteMusicState,
-    favourite_music: FavouriteMusic,
-}
-
-up!(
-    StartState -> ReceiveFullNameState,
-    ReceiveFullNameState + [full_name: String] -> ReceiveAgeState,
-    ReceiveAgeState + [age: u8] -> ReceiveFavouriteMusicState,
-    ReceiveFavouriteMusicState + [favourite_music: FavouriteMusic] -> ExitState
-);
-
-type Dialogue = Coprod!(
-    StartState,
-    ReceiveFullNameState,
-    ReceiveAgeState,
-    ReceiveFavouriteMusicState
-);
-
-wrap_dialogue!(
-    Wrapper(Dialogue),
-    default Self(Dialogue::inject(StartState))
-);
-
-// Transition functions.
-
-type Cx<State> =
-    DialogueDispatcherHandlerCx<Message, State, std::convert::Infallible>;
-type Res = ResponseResult<DialogueStage<Wrapper>>;
-
-async fn start(cx: Cx<StartState>) -> Res {
-    let DialogueDispatcherHandlerCx { cx, dialogue } = cx;
-    let dialogue = dialogue.unwrap();
-
-    cx.answer("Let's start! First, what's your full name?").send().await?;
-    next(dialogue.up())
-}
-
-async fn full_name(cx: Cx<ReceiveFullNameState>) -> Res {
-    let DialogueDispatcherHandlerCx { cx, dialogue } = cx;
-    let dialogue = dialogue.unwrap();
-
-    match cx.update.text_owned() {
-        Some(full_name) => {
-            cx.answer("What a wonderful name! Your age?").send().await?;
-            next(dialogue.up(full_name))
-        }
-        None => {
-            cx.answer("Please, send me a text message!").send().await?;
-            next(dialogue)
-        }
-    }
-}
-
-async fn age(cx: Cx<ReceiveAgeState>) -> Res {
-    let DialogueDispatcherHandlerCx { cx, dialogue } = cx;
-    let dialogue = dialogue.unwrap();
-
-    match cx.update.text().unwrap().parse() {
-        Ok(age) => {
-            cx.answer("Good. Now choose your favourite music:")
-                .reply_markup(FavouriteMusic::markup())
-                .send()
-                .await?;
-            next(dialogue.up(age))
-        }
-        Err(_) => {
-            cx.answer("Oh, please, enter a number!").send().await?;
-            next(dialogue)
-        }
-    }
-}
-
-async fn favourite_music(cx: Cx<ReceiveFavouriteMusicState>) -> Res {
-    let DialogueDispatcherHandlerCx { cx, dialogue } = cx;
-    let dialogue = dialogue.unwrap();
-
-    match cx.update.text().unwrap().parse() {
-        Ok(favourite_music) => {
-            cx.answer(format!("Fine. {}", dialogue.up(favourite_music)))
-                .send()
-                .await?;
-            exit()
-        }
-        Err(_) => {
-            cx.answer("Oh, please, enter from the keyboard!").send().await?;
-            next(dialogue)
-        }
-    }
-}
+use teloxide::prelude::*;
 
 #[tokio::main]
 async fn main() {
@@ -159,8 +47,11 @@ async fn run() {
             // Unwrap without panic because of std::convert::Infallible.
             let Wrapper(dialogue) = dialogue.unwrap();
 
-            dispatch!([cx, dialogue] -> [start, full_name, age, favourite_music])
-                .expect("Something wrong with the bot!")
+            dispatch!(
+                [cx, dialogue] ->
+                [start, receive_full_name, receive_age, receive_favourite_music]
+            )
+            .expect("Something wrong with the bot!")
         }))
         .dispatch()
         .await;
